Update JDA - xored the seed to be in line with KillTheRNG and LatiiCG - Renamed package
