+++ build.gradle
@@ -1,8 +1,9 @@
-// USE GRADLE 7
+// USE GRADLE 8
 
 plugins {
     id 'java-library'
-    id("com.github.johnrengelman.shadow") version "5.1.0"
+    id("com.github.johnrengelman.shadow") version "8.1.1"
+    id 'application'
 }
 
 sourceCompatibility = "1.8";
@@ -13,8 +14,12 @@ repositories {
     maven { url 'https://m2.dv8tion.net/releases' }
 }
 
+application {
+	mainClass='com.minecrafttas.findseedbot.main.Main'
+}
+
 dependencies {
-	implementation group: 'net.dv8tion', name: 'JDA', version: '4.3.0_348'
+	implementation group: 'net.dv8tion', name: 'JDA', version: '5.0.0-beta.9'
 	implementation group: 'ch.qos.logback', name: 'logback-classic', version: '1.2.3'
 }
